refactor all routes
