+++ backend/src/server.ts
@@ -6,10 +6,10 @@ import dotenv from 'dotenv';
 import cookieParser from 'cookie-Parser';
 dotenv.config({ path: __dirname + '/../.env' });
 import authRoutes from './routes/auth.router';
-import postRoutes from './routes/posts';
-import subRoutes from './routes/subs';
+import postRoutes from './routes/posts.router';
+import subRoutes from './routes/subs.router';
 import miscRoutes from './routes/misc.router';
-import userRoutes from './routes/user';
+import userRoutes from './routes/user.router';
 import cors from 'cors';
 import ormConfig from './ormconfig';
 //import ormConfig from './ormconfig.json';
